2023(rust): Solve Day 9: Part 1

+++ 2023/rust/days/day_09/benches/day_09_bench.rs
@@ -8,7 +8,7 @@ fn main() {
 }
 
 #[divan::bench(name = "part_one")]
-fn part_one_bench() -> Result<usize> {
+fn part_one_bench() -> Result<i32> {
     part_one(divan::black_box(&Input::from(include_str!("../input.txt"))))
 }
